build: Update the webpack files to merge common configuration.
