+++ package.json
@@ -69,8 +69,10 @@
     "style-loader": "^1.1.3",
     "url-loader": "^3.0.0",
     "webpack": "^4.42.0",
+    "webpack-sources": "^1.4.3",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3",
+    "webpack-merge": "^4.2.2",
     "whatwg-fetch": "^3.0.0"
   }
 }
